requirement.txt
tensorflow==2.14.0
numpy
pillow
matplotlib
scikit-learn
flask
opencv-python